Cambios en clientes

+++ app-customers/src/main/java/com/app/app_customers/dto/PurchaseOrderDto.java
@@ -4,6 +4,7 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -12,7 +13,8 @@ import java.util.List;
 @ToString
 public class PurchaseOrderDto {
     private String id;
-    private Integer total;
+    //private Integer total;
+    private BigDecimal total;
     private String status;
     private LocalDateTime placedOn;
     private LocalDateTime deliveredOn;
